Refactor(HmPlus): 플러그인 화
