+++ nts-admin-app/src/main.ts
@@ -6,7 +6,7 @@ import { createPinia } from 'pinia';
 
 import App from './App.vue';
 import router from './router';
-import HmPlugin from './plugins/HmPlugin';
+import HmPlus from '@/plugins/HmPlus';
 
 const pinia = createPinia();
 
@@ -15,6 +15,6 @@ const app = createApp(App);
 app.use(ElementPlus);
 app.use(pinia);
 app.use(router);
-app.use(HmPlugin);
+app.use(HmPlus);
 
 app.mount('#app');
